tests/gtestutils: temporarily disable testcases
gtestutils requires a credential file to test all features it covers.
However, the file couldn't be set up unless using personal accounts of
facebook and therefore it makes a failure[1] and affects to 'make distcheck'[2].

For sure, it must be fixed but until we find a solution, the tests are
disabled temporarily.

[1]: https://gitlab.gnome.org/GNOME/libgfbgraph/-/issues/9
[2]: https://gitlab.gnome.org/GNOME/libgfbgraph/-/issues/7